Fixed: some warnings

+++ src/game-init.cc
@@ -509,7 +509,7 @@ GameInitBox::handle_player_click(unsigned int player_index, int cx, int cy) {
   } else if ((cx > 16 + 32) && (cy < 24)) {
     if (player_index >= mission->get_player_count()) {
       mission->add_player(0, {0, 0, 0}, 20, 20, 20);
-      player_index = mission->get_player_count() - 1;
+      player_index = static_cast<unsigned int>(mission->get_player_count() - 1);
       PPlayerInfo player = mission->get_player(player_index);
       player->set_character(get_next_character(player_index));
     } else {
